Addition of seperate file for Norvig's code, all *implemented tests ok
